WordCamp Budgets: Partially remove the Due Date field.
It's no longer being used, so it can be removed everywhere. I didn't remove it from the index yet, though, because I ran into a problem with `dbDelta` updating the primary keys on the table, and fixing that isn't a blocker for the launch. We should come back and fix that next week.


git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wordcamp-qbo@2601 74240141-8908-4e6f-9713-ba540dce6ec7