WordCamp QBO: Bump class query limit to allow for growth.
The limit is just there as a sanity check, and `1000` is still low enough to serve that purpose. We'll probably have close to 150 WordCamps next year, so 300 might be too low in the next 5 years. We also already have close to 450 meetup groups, and might start creating classes for them in the foreseeable future.

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/plugins/wordcamp-qbo@5588 74240141-8908-4e6f-9713-ba540dce6ec7